requirements.txt
Flask==2.1.1
Flask_Bcrypt==1.0.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
paho_mqtt==1.6.1
WTForms==3.0.1
